refactor(core): split selection rectangle spec builder
