+++ packages/core/src/view/gridView/selectionRect.js
@@ -1,12 +1,14 @@
-import { primaryPositionalChannels } from "../../encoder/encoder.js";
 import LayerView from "../layerView.js";
 import { markGeneratedChromeOverlay } from "./generatedChromeOverlay.js";
+import {
+    createSelectionRectSpec,
+    selectionToData,
+} from "./selectionRectSpec.js";
 
-export const INTERVAL_DRAG_ACTIVE_PARAM = "intervalDragActive";
+export { INTERVAL_DRAG_ACTIVE_PARAM } from "./selectionRectSpec.js";
 
 export default class SelectionRect extends LayerView {
     /**
-     * @typedef {import("../../spec/channel.js").PrimaryPositionalChannel} PrimaryPositionalChannel
      * @typedef {import("../../types/selectionTypes.js").IntervalSelection} IntervalSelection
      */
 
@@ -19,137 +21,14 @@ export default class SelectionRect extends LayerView {
         const initialSelection = /** @type {IntervalSelection} */ (
             selectionExpr()
         );
-        const channels = Object.keys(initialSelection.intervals);
         const { zindex = 1, ...brushMarkProps } = brushConfig;
 
-        if (primaryPositionalChannels.every((c) => !channels.includes(c))) {
-            throw new Error(
-                "SelectionRect requires at least one of the channels 'x' or 'y' to be present in the selection."
-            );
-        }
-
-        /** @type {import("../../spec/view.js").LayerSpec} */
-        const layerSpec = {
-            name: "selectionRect",
-            domainInert: true,
-            params: [
-                {
-                    name: INTERVAL_DRAG_ACTIVE_PARAM,
-                    value: false,
-                },
-            ],
-            resolve: {
-                scale: {
-                    x: "forced",
-                    y: "forced",
-                },
-            },
-            data: { values: selectionToData(selectionExpr()) },
-            encoding: {},
-            layer: [],
-        };
-
-        if (channels.includes("x")) {
-            layerSpec.encoding.x = {
-                field: "_x",
-                type: null,
-                title: null,
-            };
-            layerSpec.encoding.x2 = {
-                field: "_x2",
-            };
-        }
-        if (channels.includes("y")) {
-            layerSpec.encoding.y = {
-                field: "_y",
-                type: null,
-                title: null,
-            };
-            layerSpec.encoding.y2 = {
-                field: "_y2",
-            };
-        }
-
-        layerSpec.layer.push({
-            name: "selectionRectRect",
-            mark: {
-                type: "rect",
-                clip: true,
-                ...{
-                    fill: "#808080",
-                    fillOpacity: 0.05,
-                    stroke: "black",
-                    strokeWidth: 1,
-                    strokeOpacity: 0.2,
-                    cursor: brushMarkProps.cursor ?? {
-                        expr: `${INTERVAL_DRAG_ACTIVE_PARAM} ? 'grabbing' : 'move'`,
-                    },
-                    ...brushMarkProps,
-                },
-            },
-        });
-
-        const makeExpr = (/** @type {PrimaryPositionalChannel} */ channel) => {
-            const resolution = gridChild.view.getScaleResolution(channel);
-            return (
-                `format(datum._${channel}2 - datum._${channel}, '.3s')` +
-                (resolution.type === "locus" ? " + 'b'" : "")
-            );
-        };
-
-        const labelOffset =
-            brushConfig.measure == "inside"
-                ? 9
-                : brushConfig.measure == "outside"
-                  ? -9
-                  : 0;
-
-        if (channels.includes("x") && labelOffset != 0) {
-            layerSpec.layer.push({
-                name: "selectionRectTextX",
-                mark: {
-                    type: "text",
-                    align: "center",
-                    paddingX: 5,
-                    dy: labelOffset,
-                    tooltip: null,
-                },
-                encoding: {
-                    text: { expr: makeExpr("x") },
-                    y: channels.includes("y")
-                        ? {
-                              field: "_y2",
-                              type: null,
-                              title: null,
-                          }
-                        : {
-                              value: 1,
-                          },
-                    y2: null,
-                },
-            });
-        }
-
-        if (channels.includes("y") && labelOffset != 0) {
-            layerSpec.layer.push({
-                name: "selectionRectTextY",
-                mark: {
-                    type: "text",
-                    align: "center",
-                    paddingY: 5,
-                    dy: labelOffset,
-                    tooltip: null,
-                    angle: -90,
-                },
-                encoding: {
-                    text: { expr: makeExpr("y") },
-                    x2: null,
-                },
-            });
-        }
-
         super(
-            layerSpec,
+            createSelectionRectSpec({
+                gridChild,
+                selection: initialSelection,
+                brushConfig: brushMarkProps,
+            }),
             gridChild.layoutParent.context,
             gridChild.layoutParent,
             gridChild.view,
@@ -186,33 +65,5 @@ export default class SelectionRect extends LayerView {
 
     getZindex() {
         return this._zindex;
-    }
-}
-
-/**
- *  @param {import("../../types/selectionTypes.js").IntervalSelection} selection
- */
-function selectionToData(selection) {
-    const x = selection.intervals.x;
-    const y = selection.intervals.y;
-
-    if (!x && !y) {
-        return [];
-    } else {
-        return [
-            {
-                // Using a hack here. All properties are prefixed with an underscore
-                // to prevent them from being visible in the tooltip.
-                // No properties, no tooltip. This still enables picking, masking
-                // elements under the selection rect and preventing them being
-                // selected or tooltipped.
-                // An alternative solution would necessitate adding more flags or
-                // logic to force picking in the absence of tooltips.
-                _x: x?.[0],
-                _x2: x?.[1],
-                _y: y?.[0],
-                _y2: y?.[1],
-            },
-        ];
     }
 }
